chore: ajustado subdiretório de testes do GraphQL
