+++ tests/Unit/App/GraphQL/Mutations/FundamentalMutationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Unit\GraphQL\Mutations;
+namespace Tests\Unit\App\GraphQL\Mutations;
 
 use App\GraphQL\Mutations\FundamentalMutation;
 use App\Models\Fundamental;
