+++ tests/Unit/App/GraphQL/Validators/Mutation/FundamentalCreateValidatorTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Tests\Unit\GraphQL\Validators\Mutation;
+namespace Tests\Unit\App\GraphQL\Validators\Mutation;
 
-use App\GraphQL\Validators\Mutation\PositionCreateValidator;
+use App\GraphQL\Validators\Mutation\FundamentalCreateValidator;
 use Tests\TestCase;
 
-class PositionCreateValidatorTest extends TestCase
+class FundamentalCreateValidatorTest extends TestCase
 {
     /**
      * A basic unit test messages.
@@ -14,7 +14,7 @@ class PositionCreateValidatorTest extends TestCase
      */
     public function test_messages()
     {
-        $validator = new PositionCreateValidator();
+        $validator = new FundamentalCreateValidator();
 
         $this->assertIsArray($validator->messages());
         $this->assertNotEmpty($validator->messages());
