+++ tests/Unit/App/GraphQL/Validators/Mutation/SpecificFundamentalEditValidatorTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Tests\Unit\GraphQL\Validators\Mutation;
+namespace Tests\Unit\App\GraphQL\Validators\Mutation;
 
-use App\GraphQL\Validators\Mutation\FundamentalEditValidator;
+use App\GraphQL\Validators\Mutation\SpecificFundamentalEditValidator;
 use Nuwave\Lighthouse\Execution\Arguments\ArgumentSet;
 use Tests\TestCase;
 
-class FundamentalEditValidatorTest extends TestCase
+class SpecificFundamentalEditValidatorTest extends TestCase
 {
     /**
      * A basic unit test messages.
@@ -15,7 +15,7 @@ class FundamentalEditValidatorTest extends TestCase
      */
     public function test_messages()
     {
-        $validator = new FundamentalEditValidator();
+        $validator = new SpecificFundamentalEditValidator();
 
         $this->assertIsArray($validator->messages());
         $this->assertNotEmpty($validator->messages());
@@ -31,7 +31,7 @@ class FundamentalEditValidatorTest extends TestCase
         $args = new ArgumentSet();
         $args->toArray('id');
 
-        $validator = new FundamentalEditValidator();
+        $validator = new SpecificFundamentalEditValidator();
         $validator->setArgs($args);
 
         $this->assertIsArray($validator->rules());
