+++ tests/Unit/App/GraphQL/Validators/Mutation/UserCreateValidatorTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Tests\Unit\GraphQL\Validators\Mutation;
+namespace Tests\Unit\App\GraphQL\Validators\Mutation;
 
-use App\GraphQL\Validators\Mutation\PositionEditValidator;
+use App\GraphQL\Validators\Mutation\UserCreateValidator;
 use Nuwave\Lighthouse\Execution\Arguments\ArgumentSet;
 use Tests\TestCase;
 
-class PositionEditValidatorTest extends TestCase
+class UserCreateValidatorTest extends TestCase
 {
     /**
      * A basic unit test messages.
@@ -15,7 +15,7 @@ class PositionEditValidatorTest extends TestCase
      */
     public function test_messages()
     {
-        $validator = new PositionEditValidator();
+        $validator = new UserCreateValidator();
 
         $this->assertIsArray($validator->messages());
         $this->assertNotEmpty($validator->messages());
@@ -31,7 +31,7 @@ class PositionEditValidatorTest extends TestCase
         $args = new ArgumentSet();
         $args->toArray('id');
 
-        $validator = new PositionEditValidator();
+        $validator = new UserCreateValidator();
         $validator->setArgs($args);
 
         $this->assertIsArray($validator->rules());
